Added desktop support
