+++ package.json
@@ -25,7 +25,7 @@
     "prettier": "prettier './src/**/*.{js,jsx}'",
     "prettier:check": "yarn prettier --check",
     "prettier:fix": "yarn prettier --write",
-    "start": "node scripts/start.js",
+    "start": "HTTPS=true node scripts/start.js",
     "start-mainnet": "REACT_APP_NETWORK=mainnet yarn start",
     "test": "NODE_ENV=test && node scripts/test.js --env=jsdom"
   },
@@ -50,7 +50,7 @@
     "@welldone-software/why-did-you-render": "4.0.5",
     "axios": "0.19.2",
     "bignumber.js": "9.0.0",
-    "bnc-onboard": "1.3.5",
+    "bnc-onboard": "1.3.6",
     "connected-react-router": "6.7.0",
     "currency-flags": "^2.1.1",
     "date-fns": "2.10.0",
