+++ src/components/ConnectButton/index.jsx
@@ -8,52 +8,16 @@ import { getWeb3, setWeb3 } from '~/logic/wallets/getWeb3'
 import { fetchProvider } from '~/logic/wallets/store/actions'
 import transactionDataCheck from '~/logic/wallets/transactionDataCheck'
 import { store } from '~/store'
+import { selectWallets } from '~/utils/desktop'
 
 const isMainnet = process.env.REACT_APP_NETWORK === 'mainnet'
 
 const BLOCKNATIVE_API_KEY = isMainnet ? process.env.REACT_APP_BLOCKNATIVE_KEY : '7fbb9cee-7e97-4436-8770-8b29a9a8814c'
-const PORTIS_DAPP_ID = isMainnet ? process.env.REACT_APP_PORTIS_ID : '852b763d-f28b-4463-80cb-846d7ec5806b'
-// const SQUARELINK_CLIENT_ID = isMainnet ? process.env.REACT_APP_SQUARELINK_ID : '46ce08fe50913cfa1b78'
-const FORTMATIC_API_KEY = isMainnet ? process.env.REACT_APP_FORTMATIC_KEY : 'pk_test_CAD437AA29BE0A40'
-
-const wallets = [
-  { walletName: 'metamask', preferred: true },
-  {
-    walletName: 'walletConnect',
-    preferred: true,
-    infuraKey: process.env.REACT_APP_INFURA_TOKEN,
-  },
-  {
-    walletName: 'trezor',
-    appUrl: 'gnosis-safe.io',
-    preferred: true,
-    email: 'safe@gnosis.io',
-    rpcUrl: 'https://rinkeby.infura.io/v3/b42c928da8fd4c1f90374b18aa9ac6ba',
-  },
-  {
-    walletName: 'ledger',
-    preferred: true,
-    rpcUrl: 'https://rinkeby.infura.io/v3/b42c928da8fd4c1f90374b18aa9ac6ba',
-  },
-  { walletName: 'trust', preferred: true },
-  { walletName: 'dapper' },
-  {
-    walletName: 'fortmatic',
-    apiKey: FORTMATIC_API_KEY,
-  },
-  {
-    walletName: 'portis',
-    apiKey: PORTIS_DAPP_ID,
-    label: 'Login with Email',
-  },
-  { walletName: 'authereum' },
-  { walletName: 'coinbase' },
-  { walletName: 'opera' },
-  { walletName: 'operaTouch' },
-]
 
 let lastUsedAddress = ''
 let providerName
+
+const wallets = selectWallets()
 
 export const onboard = new Onboard({
   dappId: BLOCKNATIVE_API_KEY,
